+++ engine/standalone/engine/src/test/scala/pl/touk/nussknacker/engine/standalone/test/StandaloneTestMainSpec.scala
@@ -25,6 +25,7 @@ class StandaloneTestMainSpec extends FunSuite with Matchers with BeforeAndAfterE
   test("perform test on mocks") {
     val process = EspProcessBuilder
       .id("proc1")
+      .path(None)
       .exceptionHandler()
       .source("start", "request1-post-source")
       .filter("filter1", "#input.field1() == 'a'")
@@ -65,6 +66,7 @@ class StandaloneTestMainSpec extends FunSuite with Matchers with BeforeAndAfterE
   test("detect errors in nodes") {
     val process = EspProcessBuilder
       .id("proc1")
+      .path(None)
       .exceptionHandler()
       .source("start", "request1-post-source")
       .filter("occasionallyThrowFilter", "#input.field1() == 'a' ? 1/0 == 0 : true")
@@ -93,6 +95,7 @@ class StandaloneTestMainSpec extends FunSuite with Matchers with BeforeAndAfterE
   test("get results on parameter sinks") {
     val process = EspProcessBuilder
       .id("proc1")
+      .path(None)
       .exceptionHandler()
       .source("start", "request1-post-source")
       .emptySink("endNodeIID", "parameterResponse-sink", "computed" -> "#input.field1()")